fixed release build workflow (#118)
