+++ .github/workflows/release.yml
@@ -45,11 +45,30 @@ jobs:
         with:
           ref: ${{ steps.git-tag.outputs.git-version }}
 
+      - name: Get npm version
+        id: npm-tag
+        uses: martinbeentjes/npm-get-version-action@master
+
       - name: Output versions
-        run: echo building relase from git ${{ steps.git-tag.outputs.git-version }}
+        run: |
+          echo git ${{ steps.git-tag.outputs.git-version }}
+          echo npm ${{ steps.npm-tag.outputs.current-version }}
+
+      - name: Versions not matching
+        if: steps.git-tag.outputs.git-version != steps.npm-tag.outputs.current-version
+        run: |
+          echo git and npm versions not equal - refusing to build release
+          exit 1
+
+      - name: Versions match
+        run: |
+          echo versions equal - building release ${{ steps.git-tag.outputs.git-version }}
 
       - name: Install Dependencies
         run: yarn
+
+      - name: Create release version metadata
+        run: yarn build:release ${{ steps.git-tag.outputs.git-version }}
 
       - name: Build Application
         run: yarn build
