+++ public/documentation/data/Releases.json
@@ -1,11 +1,3 @@
 [
-  "refs/tags/1.0.0",
-  "refs/tags/1.0.0-RC8",
-  "refs/tags/1.0.0-RC7",
-  "refs/tags/1.0.0-RC6",
-  "refs/tags/1.0.0-RC5",
-  "refs/tags/1.0.0-RC4",
-  "refs/tags/1.0.0-RC3",
-  "refs/tags/1.0.0-RC2",
-  "refs/tags/1.0.0-RC1"
+  "main"
 ]
